+++ public/CHANGELOG.md
@@ -19,6 +19,7 @@ v1.2.18
 - Added copy and delete buttons for each entry in the address book.
 - Integrated the address book into the send modal for quick recipient selection.
 - Integrated the address book into the exchange modal (Sell STH) for quick USDT address selection.
+- Added a check for the maximum STH buy amount based on the hot wallet balance.
 
 v1.2.17
 - Add support 7 languages